LolKing has been shutdown so support has been removed.
Solomid.net has not been updated or touched in 2 years so it has also been removed as the guides are not only extremely outdated but many of the items mentioned are drastically changed or non-existent as well as many champions reworked.

Updated the ReadMe & Manifest as well to reflect new versions + the feature changes.